Continuing bug fixes in SubstitutionTreeSearch.
It seems as if getting the full tree search may fail at the first hurdle
as the lambda coefficients are simply not being generated correctly at
all. This would be fixed with access to a sufficiently large word corpus
and n-gram list, but as this is unlikely so too seems fixing the issues
currently seen.